+++ homeassistant/util/aiohttp.py
@@ -3,7 +3,6 @@ import json
 from urllib.parse import parse_qsl
 from typing import Any, Dict, Optional
 
-from aiohttp import web
 from multidict import CIMultiDict, MultiDict
 
 
@@ -42,12 +41,3 @@ class MockRequest:
     async def text(self) -> str:
         """Return the body as text."""
         return self._text
-
-
-def serialize_response(response: web.Response) -> Dict[str, Any]:
-    """Serialize an aiohttp response to a dictionary."""
-    return {
-        'status': response.status,
-        'body': response.text,
-        'headers': dict(response.headers),
-    }
